LPS-87431 Refactor
