+++ modules/apps/segments/segments-service/src/test/resources/com/liferay/segments/internal/asah/client/data/binding/get-individual-segments.json
